requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai==1.14.0
groq==1.1.2
numpy==2.2.3
pymupdf==1.25.3
python-dotenv==1.1.0
